pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "isp-uptime-monitor"
version = "0.1.0"
description = "ISP Uptime Monitoring & Speed Testing (lightweight)"
authors = [
    { name="Josh Whitcomb", email="boostedviii@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyQt6>=6.6.0",
    "pyqtgraph>=0.13.3",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ping3>=4.0.0",
    "speedtest-cli>=2.1.3"
]

[project.scripts]
isp-uptime-monitor = "isp_monitor.main:main" 